Define new project structure
