--- notebooks/context.py
@@ -1,10 +1,0 @@
-"""
-Thanks to this:
-https://docs.python-guide.org/writing/structure/#test-suite
-"""
-
-import os
-import sys
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-
-import perceptrons
