--- perceptrons/activations.py
@@ -1,7 +1,0 @@
-import numpy as np
-
-
-def heaviside(z):
-    """Heaviside step function."""
-    a = np.where(z >= 0, 1, 0)
-    return a
